added controllers for previous tasks
