+++ routes/eraseEvents.js
@@ -1,8 +1,10 @@
 var express = require('express');
 var router = express.Router();
 
-// Route related to delete events
 
-router.get('/')
+var eventController = require('../controllers/events')
+
+
+router.delete('/', eventController.eraseEvents)
 
 module.exports = router;
